+++ Sources/ParseSwift/Coding/ObjectType.swift
@@ -8,70 +8,9 @@
 
 import Foundation
 
-public struct NoBody: Codable {}
-
-public protocol Saving: Codable {
-    associatedtype SavingType
-    func save(options: API.Options) throws -> SavingType
-    func save() throws -> SavingType
-}
-
-extension Saving {
-    public func save() throws -> SavingType {
-        return try save(options: [])
-    }
-}
-
-public protocol Fetching: Codable {
-    associatedtype FetchingType
-    func fetch(options: API.Options) throws -> FetchingType
-    func fetch() throws -> FetchingType
-}
-
-extension Fetching {
-    public func fetch() throws -> FetchingType {
-        return try fetch(options: [])
-    }
-}
-
-public protocol ObjectType: Fetching, Saving, CustomDebugStringConvertible, Equatable {
-    static var className: String { get }
-    var objectId: String? { get set }
-    var createdAt: Date? { get set }
-    var updatedAt: Date? { get set }
-    var ACL: ACL? { get set }
-}
-
 internal extension ObjectType {
     func getEncoder() -> ParseEncoder {
         return getParseEncoder()
-    }
-}
-
-extension ObjectType {
-    // Parse ClassName inference
-    public static var className: String {
-        let classType = "\(type(of: self))"
-        return classType.components(separatedBy: ".").first! // strip .Type
-    }
-    public var className: String {
-        return Self.className
-    }
-}
-
-extension ObjectType {
-    public var debugDescription: String {
-        guard let descriptionData = try? getJSONEncoder().encode(self),
-            let descriptionString = String(data: descriptionData, encoding: .utf8) else {
-                return "\(className) ()"
-        }
-        return "\(className) (\(descriptionString))"
-    }
-}
-
-public extension ObjectType {
-    func toPointer() -> Pointer<Self> {
-        return Pointer(self)
     }
 }
 
@@ -158,46 +97,4 @@ func getDecoder() -> JSONDecoder {
     let encoder = JSONDecoder()
     encoder.dateDecodingStrategy = dateDecodingStrategy
     return encoder
-}
-
-public extension ObjectType {
-    func save(options: API.Options) throws -> Self {
-        return try saveCommand().execute(options: options)
-    }
-
-    func fetch(options: API.Options) throws -> Self {
-        return try fetchCommand().execute(options: options)
-    }
-
-    internal func saveCommand() -> API.Command<Self, Self> {
-        return API.Command<Self, Self>.saveCommand(self)
-    }
-
-    internal func fetchCommand() throws -> API.Command<Self, Self> {
-        return try API.Command<Self, Self>.fetchCommand(self)
-    }
-}
-
-extension ObjectType {
-    var endpoint: API.Endpoint {
-        if let objectId = objectId {
-            return .object(className: className, objectId: objectId)
-        }
-        return .objects(className: className)
-    }
-
-    var isSaved: Bool {
-        return objectId != nil
-    }
-}
-
-public struct FindResult<T>: Decodable where T: ObjectType {
-    let results: [T]
-    let count: Int?
-}
-
-public extension ObjectType {
-    var mutationContainer: ParseMutationContainer<Self> {
-        return ParseMutationContainer(target: self)
-    }
 }
