+++ Sources/ParseSwift/Parse Types/NoBody.swift
@@ -1,0 +1,9 @@
+//
+//  NoBody.swift
+//  ParseSwift
+//
+//  Created by Pranjal Satija on 7/18/20.
+//  Copyright © 2020 Parse. All rights reserved.
+//
+
+public struct NoBody: Codable {}
